+++ .storybook/main.js
@@ -3,11 +3,13 @@ const tsconfigPaths = require('vite-tsconfig-paths').default;
 module.exports = {
   stories: ['../packages/**/*.story.ts', '../stories/**/*.story.ts'],
   addons: [
-    '@storybook/addon-links',
     '@storybook/addon-essentials',
+    '@storybook/addon-links',
     '@storybook/addon-a11y',
     '../storyhelpers/storybook-readme/preset.js',
   ],
+  framework: '@storybook/web-components',
+  features: { storyStoreV7: false },
   core: { builder: '@storybook/builder-vite' },
   staticDirs: ['./images'],
 
@@ -22,16 +24,6 @@ module.exports = {
     if (configType === 'DEVELOPMENT') {
       // add plugins
       config.plugins.push(tsconfigPaths());
-
-      // pre bundle dependencies for faster startup time
-      config.optimizeDeps.include.push('@storybook/web-components');
-      config.optimizeDeps.include.push('lit-html');
-      config.optimizeDeps.include.push('lit');
-      config.optimizeDeps.include.push('lit/decorators.js');
-      config.optimizeDeps.include.push('lit/directives/style-map.js');
-      config.optimizeDeps.include.push('lit/directives/if-defined.js');
-      config.optimizeDeps.include.push('lit/directives/unsafe-html.js');
-      config.optimizeDeps.include.push('element-internals-polyfill');
     }
 
     return config;